+++ CHANGELOG.md
@@ -4,8 +4,14 @@
 
 ## [Unreleased]
 
+### Added
+
+- New Icons
+
 ## [0.0.2]
+
 ### Implemented
+
 - Load from json, Show In Tool Pane, Navigate with click
 - Navigate with Previous/Next buttons on Tool Pane
 - Create|Delete Tour from Tool Pane